Refactor PiecewiseBase (#14220)
