+++ framework/include/functions/PiecewiseBase.h
@@ -30,20 +30,37 @@ public:
 
   PiecewiseBase(const InputParameters & parameters);
 
-  virtual void initialSetup();
   virtual Real functionSize() const;
   virtual Real domain(const int i) const;
   virtual Real range(const int i) const;
 
   /**
    * Provides a means for explicitly setting the x and y data. This must
-   * be called in the constructor of inhereted classes.
+   * be called in the constructor of inherited classes.
    */
   void setData(const std::vector<Real> & x, const std::vector<Real> & y);
 
 protected:
-  std::unique_ptr<LinearInterpolation> _linear_interp;
+  ///@{ raw function data as read
+  std::vector<Real> _raw_x;
+  std::vector<Real> _raw_y;
+  ///@}
+
+  /// function value scale factor
+  const Real _scale_factor;
+
+  ///@{ if _has_axis is true point component to use as function argument, otherwise use t
   int _axis;
   const bool _has_axis;
-  bool _data_set;
+  ///@}
+
+private:
+  /// Reads data from supplied CSV file.
+  void buildFromFile();
+
+  /// Builds data from 'x' and 'y' parameters.
+  void buildFromXandY();
+
+  /// Builds data from 'xy_data' parameter.
+  void buildFromXY();
 };
